backend/requirements-dev.txt
# Testing dependencies (pytest, pytest-cov, pytest-asyncio already in requirements-core.txt)
pytest-mock==3.12.0
pytest-env==1.1.3
httpx==0.28.1  # For testing HTTP clients

# Mocking and fixtures
faker==20.1.0
factory-boy==3.3.0
responses==0.24.1
httpretty==1.1.7

# Code quality
black==23.12.1
ruff==0.1.8
mypy==1.7.1

# Coverage reporting
coverage[toml]==7.3.3
